+++ CFE_GestionRecibos/Empleado/Empleado.cs
@@ -12,7 +12,7 @@ namespace CFE_GestionRecibos.Empleado
 {
     public partial class Empleado : Form
     {
-        public int id = 001;
+        public long id = 001;
         public string username = "Default Employee";
 
         public Empleado()
